hyperv: move copyFile to the crc/os package
This will allow to share some of its code with a CopySparse method which
will be introduced in a follow-up commit.
